Rename services field to collectors
The `collectors` field will contain a list of collectors from an agent. The `services` field was supposed to do that, but the name `collectors` is more fitting and won't conflict with other BE fields.

From the agent side the `services` field was always set to NULL and never had anything transmitted over it using the new architecture.