Remove unused import in proto/agent/v1/cmds.proto
Per the protobuf compiler output, this import is unused in this file. This results in a warning being issued whenever it’s compiled. Comment it out for now to resolve that warning.

This is part of an ongoing project I’m working on to eliminate warnings from our build output as much as possible (and thus make it easier to find build errors in build output in CI).